+++ CHANGES.md
@@ -7,6 +7,8 @@ ChangeLog
 Bug Handling
 ************
 - don't allow running edit as root
+- default to /usr/bin/vi for edit
+- check that the provided editor is executable
 
 Features
 ********